Create Product Landing page HTML
This is for the product landing page. It meets all requirements. But the page is limited to some features and doesnt look as good as I had it at one point. But I still think its pretty good considering I had to change some things around to meet the tests requirements. Also links and stuff are for the user to input. This is a fake product made by me for this tests so i couldnt add any legit videos or links with photos.